requirements.txt
Flask==3.0.2
pytest==8.0.1
python-dotenv==1.0.1
pytest-cov==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
gunicorn==21.2.0
Flask-WTF==1.2.1
email_validator==2.1.1
